Move StatusIcon to components/common
