+++ airbyte-webapp/src/components/EntityTable/components/NameCell.tsx
@@ -3,8 +3,8 @@ import { useIntl } from "react-intl";
 import styled from "styled-components";
 
 import { ConnectorIcon } from "components/common/ConnectorIcon";
-import StatusIcon from "components/StatusIcon";
-import { StatusIconStatus } from "components/StatusIcon/StatusIcon";
+import { StatusIcon } from "components/common/StatusIcon";
+import { StatusIconStatus } from "components/common/StatusIcon/StatusIcon";
 
 import { Status } from "../types";
