--- airbyte-webapp/src/components/StatusIcon/index.tsx
@@ -1,4 +1,0 @@
-import StatusIcon from "./StatusIcon";
-
-export default StatusIcon;
-export { StatusIcon };
