+++ airbyte-webapp/src/components/common/StatusIcon/CircleLoader.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/colors";
+@use "scss/colors";
 
 @keyframes spinning {
   0% {
