+++ airbyte-webapp/src/components/common/StatusIcon/index.stories.tsx
@@ -1,0 +1,18 @@
+import { ComponentStory, ComponentMeta } from "@storybook/react";
+
+import { StatusIcon } from "./StatusIcon";
+
+export default {
+  title: "common/StatusIcon",
+  component: StatusIcon,
+  argTypes: {
+    value: { type: { name: "number", required: false } },
+  },
+} as ComponentMeta<typeof StatusIcon>;
+
+const Template: ComponentStory<typeof StatusIcon> = (args) => <StatusIcon {...args} />;
+
+export const Primary = Template.bind({});
+Primary.args = {
+  status: "success",
+};
