+++ airbyte-webapp/src/components/common/StatusIcon/index.ts
@@ -1,0 +1,1 @@
+export * from "./StatusIcon";
