+++ airbyte-webapp/src/pages/SettingsPage/pages/ConnectorsPage/components/CreateConnectorModal.tsx
@@ -4,7 +4,8 @@ import { FormattedMessage, useIntl } from "react-intl";
 import styled from "styled-components";
 import * as yup from "yup";
 
-import { LabeledInput, Link, StatusIcon } from "components";
+import { LabeledInput, Link } from "components";
+import { StatusIcon } from "components/common/StatusIcon";
 import { Button } from "components/ui/Button";
 import { Modal } from "components/ui/Modal";
