+++ airbyte-webapp/src/views/Connector/ServiceForm/components/TestingConnectionError.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { FormattedMessage } from "react-intl";
 import styled from "styled-components";
 
-import { StatusIcon } from "components";
+import { StatusIcon } from "components/common/StatusIcon";
 
 const Error = styled(StatusIcon)`
   padding-left: 1px;
